Enabling building shared libraries.
This commit allows the ability to build shared libraries using
-DBUILD_SHARED_LIBS=ON at the cmake configuration. This is turned on by
default.

The logic that has been added to the cmake script provides the ability
to build both static and shared libs in the same build process by first
building an object library.

The unit testing is using the object library for testing.

This commit references NOAA-EMC/NCEPLIBS-sp#67